% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_add_pointID.R
\name{add_pointid_col}
\alias{add_pointid_col}
\title{add same id to same point}
\usage{
add_pointid_col(df)
}
\arguments{
\item{df}{kouji dataframe}
}
\value{
data.frame
}
\description{
add same id to same point
}
