% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_koujiID.R
\name{make_koujiID}
\alias{make_koujiID}
\title{Make string for koujiID}
\usage{
make_koujiID(youto, renban, name)
}
\arguments{
\item{youto}{digit string 0 or 5 or 7 etc}

\item{renban}{digit string for serial number}

\item{name}{string for name of city or town etc}
}
\value{
char
}
\description{
Make string for koujiID which is Hyoujyuntachibabgou
by some element.
}
