% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_serial_price.R
\name{make_serial_price_data}
\alias{make_serial_price_data}
\title{Making serial price Rdata}
\usage{
make_serial_price_data(csvfile)
}
\arguments{
\item{csvfile}{}
}
\value{
df
}
\description{
Making serial price Rdata
}
