% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_collecting_data.R
\name{pickup_from_zip}
\alias{pickup_from_zip}
\title{pick up specific file from zip archive}
\usage{
pickup_from_zip(zipfiles, pattern = "", dest_dir = ".")
}
\arguments{
\item{dest_dir}{}
}
\description{
pick up specific file from zip archive
}
