% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_readtool.R
\name{read_kouji_csv}
\alias{read_kouji_csv}
\title{make R data by reading Chikakouji csv file}
\usage{
read_kouji_csv(csvfile)
}
\arguments{
\item{csvfile}{paths of Chikakouji csv files}
}
\value{
dataframe
}
\description{
make R data by reading Chikakouji csv file
}
