% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_collecting_data.R
\name{url_list_kouji}
\alias{url_list_kouji}
\title{url list of a prefecture}
\usage{
url_list_kouji(pref = 27)
}
\arguments{
\item{pref}{prefecuture id}
}
\value{
data.frame
}
\description{
url list of a prefecture
}
